kernel: support splitted sdu_write hack
This is a temporary hack that makes it possible to port TCP
applications without worrying about the size of write()/send()
system calls, which may well exceed the MTU size.
For those applications, rl_io_sdu_write() can split of the user
buffer in multiple SDUs, since message boundaries are not
required. The application must activate the hack (for each I/O
device) after flow_alloc() or flow_accept(), by means of an
ioctl(). Once fragmentation and reassenbly is supported, the
ioctl() will be changed so not to have any effect, so that it
will not even necessary to recompile the applications.